+++ Backend/src/controller/authController/authController.ts
@@ -1,0 +1,2 @@
+export { default as loginUser } from "./loginUser";
+export { default as registerUser } from "./registerUser";
