+++ Backend/src/server.ts
@@ -1,10 +1,9 @@
 import express, { Request, Response, NextFunction } from "express";
 import cors from "cors";
 import dotenv from "dotenv";
-import { connectDB } from "../db/connectDB";
-
-import todosRouter from "../router/todosRouter";
-import authRouter from "../router/authRouter";
+import { connectDB } from "./db/connectDB";
+import todosRouter from "./router/todosRouter";
+import authRouter from "./router/authRouter";
 
 dotenv.config();
 
@@ -16,6 +15,8 @@ app.use(todosRouter);
 
 connectDB();
 
-app.listen(5000, () => {
-  console.log("Server is running on port 5000");
+const PORT = process.env.PORT;
+
+app.listen(PORT, () => {
+  console.log(`Server is running on port ${PORT}`);
 });
