+++ Frontend/src/interface/Todo.ts
@@ -1,0 +1,5 @@
+export interface Todo {
+  id: number;
+  task: string;
+  completed: boolean;
+}
