+++ Frontend/src/pages/Register.tsx
@@ -1,45 +1,13 @@
 import React, { useState } from "react";
-import axios from "axios";
-import Swal from "sweetalert2";
 import { useNavigate } from "react-router-dom";
 import { Container, Row, Col, Form, Button } from "react-bootstrap";
 import "bootstrap/dist/css/bootstrap.min.css";
+import { handleRegister } from "../services/authServices/authServices";
 
 const Register: React.FC = () => {
   const [username, setUsername] = useState<string>("");
   const [password, setPassword] = useState<string>("");
   const navigate = useNavigate();
-
-  const handleRegister = async () => {
-    try {
-      await axios.post("http://localhost:5000/register", {
-        username,
-        password,
-      });
-      Swal.fire({
-        icon: "success",
-        title: "Registration Successful",
-        text: "You can now log in!",
-        background: "#d4edda",
-        color: "#155724",
-        confirmButtonColor: "#3085d6",
-        confirmButtonText: "Login",
-      }).then(() => {
-        navigate("/login");
-      });
-    } catch (err: any) {
-      Swal.fire({
-        icon: "error",
-        title: "Registration Failed",
-        text:
-          err.response?.data.message || "An error occurred. Please try again.",
-        background: "#f8d7da",
-        color: "#721c24",
-        confirmButtonColor: "#3085d6",
-        confirmButtonText: "Try Again",
-      });
-    }
-  };
 
   return (
     <Container
@@ -73,7 +41,7 @@ const Register: React.FC = () => {
             <Button
               className="mt-4 w-100"
               variant="primary"
-              onClick={handleRegister}
+              onClick={() => handleRegister(username, password, navigate)}
             >
               Register
             </Button>
