+++ Frontend/src/services/authServices/authServices.ts
@@ -1,0 +1,3 @@
+export { default as handleLogin } from "./handleLogin";
+export { default as handleLogout } from "./handleLogout";
+export { default as handleRegister } from "./handleRegister";
