+++ Frontend/src/services/todoServices/todoServices.ts
@@ -1,0 +1,5 @@
+export { default as handleAddTask } from "../todoServices/handleAddTask";
+export { default as handleDeleteTask } from "../todoServices/handleDeleteTask";
+export { default as handleEditTask } from "../todoServices/handleEditTask";
+export { default as handleToggleCompleted } from "../todoServices/handleToggleCompleted";
+export { default as handleUpdateTask } from "../todoServices/handleUpdateTask";
